Fix annoying warnings

+++ Eigen/src/Core/Ref.h
@@ -334,14 +334,14 @@ template<typename TPlainObjectType, int Options, typename StrideType> class Ref<
     typedef internal::traits<Ref> Traits;
 
     static constexpr bool may_map_m_object_successfully = 
-      (StrideType::InnerStrideAtCompileTime == 0 ||
-       StrideType::InnerStrideAtCompileTime == 1 ||
-       StrideType::InnerStrideAtCompileTime == Dynamic) &&
+      (static_cast<int>(StrideType::InnerStrideAtCompileTime) == 0 ||
+       static_cast<int>(StrideType::InnerStrideAtCompileTime) == 1 ||
+       static_cast<int>(StrideType::InnerStrideAtCompileTime) == Dynamic) &&
       (TPlainObjectType::IsVectorAtCompileTime ||
-       StrideType::OuterStrideAtCompileTime == 0 ||
-       StrideType::OuterStrideAtCompileTime == Dynamic ||
-       StrideType::OuterStrideAtCompileTime == TPlainObjectType::InnerSizeAtCompileTime ||
-       TPlainObjectType::InnerSizeAtCompileTime == Dynamic);
+       static_cast<int>(StrideType::OuterStrideAtCompileTime) == 0 ||
+       static_cast<int>(StrideType::OuterStrideAtCompileTime) == Dynamic ||
+       static_cast<int>(StrideType::OuterStrideAtCompileTime) == static_cast<int>(TPlainObjectType::InnerSizeAtCompileTime) ||
+       static_cast<int>(TPlainObjectType::InnerSizeAtCompileTime) == Dynamic);
   public:
 
     typedef RefBase<Ref> Base;
